Restructure nav to allow dynamic menu
