+++ app/javascript/components/navigation/index.js
@@ -1,8 +1,11 @@
 import React, { Component, Fragment as F } from "react";
 import { Container } from "semantic-ui-react";
+import PropTypes from "prop-types";
+
 import NavBar from "./responsiveNavbar";
 
-import links from "./navlinks";
+import authorized from "./links/auth";
+import base from "./links/base";
 import "./nav.scss";
 
 class Nav extends Component {
@@ -10,17 +13,27 @@ class Nav extends Component {
         super(props);
         //eslint-disable-next-line
         console.log(props);
-        this.state = {};
+        const { loggedIn } = props;
+        this.state = {
+            links: loggedIn ? authorized : base
+        };
     }
 
     render() {
+        const { links } = this.state;
         return (
             <F>
                 <NavBar leftItems={links.left} rightItems={links.right} />
-                <Container style={{ height: 1000, background: "white" }}>This is some content</Container>
+                <div className="navbar-spacer" style={{ marginTop: "68px" }} />
+                <Container style={{ height: 100, background: "white" }}>This is some spacer content</Container>
             </F>
         );
     }
 }
+
+const { bool } = PropTypes;
+Nav.propTypes = {
+    loggedIn: bool.isRequired
+};
 
 export default Nav;
