+++ app/javascript/components/navigation/links/base.js
@@ -1,0 +1,10 @@
+const links = {
+    left: [],
+    right: [
+        { as: "a", content: "About", key: "about", href: "/about" },
+        { as: "a", content: "Contact", key: "contact", href: "/contact" },
+        { as: "a", content: "Register/Login", key: "register", href: "/users/sign_in" }
+    ]
+};
+
+export default links;
